+++ Example/CryptoOnramp Example/CryptoOnramp Example/RegistrationView.swift
@@ -51,10 +51,10 @@ struct RegistrationView: View {
     }
 
     private var isUpdatePhoneNumberButtonDisabled: Bool {
-        !isRegistrationComplete
+        isLoading.wrappedValue || !isRegistrationComplete
     }
 
-    private var shouldDisableButtons: Bool {
+    private var isAuthenticateButtonDisabled: Bool {
         isLoading.wrappedValue
     }
 
@@ -102,15 +102,17 @@ struct RegistrationView: View {
 
                 if isRegistrationComplete {
                     Button("Authenticate") {
+                        resetFocusState()
                         Task {
                             try await verify()
                         }
                     }
                     .buttonStyle(PrimaryButtonStyle())
-                    .disabled(shouldDisableButtons)
-                    .opacity(shouldDisableButtons ? 0.5 : 1)
+                    .disabled(isAuthenticateButtonDisabled)
+                    .opacity(isAuthenticateButtonDisabled ? 0.5 : 1)
                 } else {
                     Button("Register") {
+                        resetFocusState()
                         registerUser()
                     }
                     .buttonStyle(PrimaryButtonStyle())
@@ -119,6 +121,7 @@ struct RegistrationView: View {
                 }
 
                 Button("Update Phone Number") {
+                    resetFocusState()
                     updatePhoneNumberInput = phoneNumber
                     showUpdatePhoneNumberSheet = true
                 }
@@ -254,6 +257,12 @@ struct RegistrationView: View {
                 }
             }
         }
+    }
+
+    private func resetFocusState() {
+        isFullNameFieldFocused = false
+        isPhoneNumberFieldFocused = false
+        isCountryFieldFocused = false
     }
 }