+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## XX.YY.ZZ 2025-xx-yy
+## 24.23.2 2025-09-15
 ## PaymentSheet
 * [Fixed] Fixed an issue with the Liquid Glass keyboard toolbar where tapping on the sheet immediately above the keyboard didn't have an effect.